03-ai-products/svo-payments-bot/requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
aiosqlite==0.20.0
pytest==8.3.4
pytest-asyncio==0.25.0
